# backend/openapi-reservation-fullstack.yaml
openapi: 3.0.3
info:
  title: API Réservation de ressources
  version: 2.0.0
  description: >
    Contrat Frontend/Backend pour une plateforme de réservation en ligne sécurisée.
  contact:
    name: Yuliia Kolomiiets
servers:
  - url: /api
    description: Serveur API
security:
  - BearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getMyProfile
      summary: Consulter mon profil utilisateur
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    
    patch:
      tags: [Users]
      operationId: updateMyProfile
      summary: Mettre à jour mon profil
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    
    delete:
      tags: [Users]
      operationId: deleteMyAccount
      summary: Supprimer mon compte utilisateur
      description: >
        Supprime le compte de l'utilisateur connecté.
        Envoie une notification email de confirmation de suppression.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Compte supprimé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Compte supprimé avec succès"
                  emailSent:
                    type: boolean
                    description: Indique si l'email de confirmation a été envoyé
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===== RESOURCES =====
  /resources:
    get:
      tags: [Resources]
      operationId: listResources
      summary: Lister les ressources disponibles
      description: >
        Retourne la liste paginée des ressources correspondant aux filtres.
        La liste peut être vide (cas normal).
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/capacity"
        - $ref: "#/components/parameters/location"
        - $ref: "#/components/parameters/amenities"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Succès (liste possiblement vide)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResources"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    
    post:
      tags: [Resources]
      operationId: createResource
      summary: Créer une nouvelle ressource (admin uniquement)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceRequest"
      responses:
        "201":
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /resources/{id}:
    get:
      tags: [Resources]
      operationId: getResource
      summary: Récupérer le détail d'une ressource
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    
    patch:
      tags: [Resources]
      operationId: updateResource
      summary: Modifier une ressource (admin uniquement)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceRequest"
      responses:
        "200":
          description: Ressource modifiée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    
    delete:
      tags: [Resources]
      operationId: deleteResource
      summary: Supprimer/désactiver une ressource (admin uniquement)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Ressource supprimée/désactivée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ressource désactivée avec succès"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT fourni par Clerk
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      description: Identifiant de la ressource
      schema:
        type: string
        example: resource_123
...
    start:
      name: start
      in: query
      required: true
      description: Début du créneau (ISO 8601)
      schema:
        type: string
        format: date-time
        example: "2026-01-19T09:00:00Z"
    end:
      name: end
      in: query
      required: true
      description: Fin du créneau (ISO 8601)
      schema:
        type: string
        format: date-time
        example: "2026-01-19T10:00:00Z"
...
    capacity:
      name: capacity
      in: query
      required: false
      description: Capacité minimale (nombre de personnes)
      schema:
        type: integer
        minimum: 1
        example: 6
    location:
      name: location
      in: query
      required: false
      description: Filtre par localisation (texte libre)
      schema:
        type: string
        example: "Paris 10"
    amenities:
      name: amenities
      in: query
      required: false
      description: Filtre équipements sous forme CSV (ex wifi,ecran)
      schema:
        type: string
        example: "wifi,ecran"
    page:
      name: page
      in: query
      required: false
      description: Numéro de page (indexé à 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      required: false
      description: Taille de page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    Unauthorized:
      description: Non authentifié - token manquant ou invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Token d'authentification manquant ou invalide"
    Forbidden:
      description: Accès interdit - permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "FORBIDDEN"
              message: "Vous n'avez pas les permissions nécessaires pour cette action"
    BadRequest:
      description: Paramètres ou données invalides
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalServerError:
      description: Erreur technique
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    User:
      type: object
      additionalProperties: false
      required: [id, email, fullName, role, createdAt]
      properties:
        id:
          type: string
          description: Identifiant unique de l'utilisateur
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          type: string
          enum: [user, admin]
          description: Rôle de l'utilisateur
        phone:
          type: string
          nullable: true
        twoFactorEnabled:
          type: boolean
          description: Indique si la 2FA est activée
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      additionalProperties: false
      properties:
        fullName:
          type: string
        phone:
          type: string
        twoFactorEnabled:
          type: boolean

    # ===== RESOURCES =====
    Resource:
      type: object
      additionalProperties: false
      required: [id, name, description, active, availability, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
          minimum: 1
          nullable: true
          description: Capacité maximale (optionnel)
        location:
          type: string
          nullable: true
        amenities:
          type: array
          items:
            type: string
          default: []
        pricePerHour:
          type: number
          minimum: 0
          nullable: true
        images:
          type: array
          items:
            type: string
            format: uri
          default: []
        active:
          type: boolean
          description: Indique si la ressource est activée/disponible
        availability:
          type: object
          description: Règles de disponibilité (jours/horaires)
          properties:
            daysOfWeek:
              type: array
              items:
                type: integer
                minimum: 0
                maximum: 6
              description: Jours de la semaine (0=dimanche, 6=samedi)
            timeRanges:
              type: array
              items:
                type: object
                properties:
                  start:
                    type: string
                    pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
                    example: "09:00"
                  end:
                    type: string
                    pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
                    example: "18:00"
        rules:
          type: object
          description: Règles spécifiques
          properties:
            minDurationMinutes:
              type: integer
              minimum: 1
            maxDurationMinutes:
              type: integer
              minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedResources:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    CreateResourceRequest:
      type: object
      additionalProperties: false
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
          minimum: 1
        location:
          type: string
        amenities:
          type: array
          items:
            type: string
        pricePerHour:
          type: number
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
        availability:
          type: object
          description: Règles de disponibilité
        rules:
          type: object
          description: Règles spécifiques
    UpdateResourceRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
          minimum: 1
        location:
          type: string
        amenities:
          type: array
          items:
            type: string
        pricePerHour:
          type: number
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
        active:
          type: boolean
        availability:
          type: object
        rules:
          type: object

    # ===== AVAILABILITY =====
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              default: {}